Add support for SNURL passive mode
This is especially interesting with multicast-only transmitters. It
allows snooping on the data without interfering with the stream, while a
proper receiver can still process all the data without issues.

This may case the passive snooper to not receive all data, because it
might lose a packet with the only copy of a frame which has been
received by the primary recipient.