Implement "streamify" node
This node takes a series of sample values and puts them through the
stream buffer infrastructure. This effectively converts subsequent
samples into a stream, which can then be used in the stream processing
nodes (i.e. summary, detrend, fft).

The reason we do not do this by default is that explicit configuration
of the slicing and period of the data is required in order to properly
streamify it. In addition, streams do not support components, which
means that for a single device, only a single stream can be reasonably
generated.