Decoder for BME68x data
This decoder uses floating-point math instead of fixed-point math as we
did for the BME280. While floating point math is more expensive, the
fixed point implementation of the pressure decoding in the datasheet has
significantly lower precision; simple attempts to raise the precision
have resulted in overflows.

Because I already fought with this way too long, we'll be going with
floating point here.